Prepare zb_sum to be incorporated into Sum
Changed the arguments so that the function can be called just with
what is provided to sum. For now we will check conditions so that
vaniish is true and won't attempt to use zb_sum if it is not.

Also corrected some limits that would be wrong for more exotic
upper limits than n + c.