Add function to detect if a function vanishes
Now zb_sum can be called with just the information provided to it
by summation. So it can now conceptually be included. There is now
a function which tries to work out if a hypergeometric term vanishes
past some window by looking at the roots of F(n, k + 1) / F(n, k).

Noticed that thing were being called hypergeometric functions when
what was meant was hypergeometric term, so that's changed.