Integrate zb_sum into sum
Now summation calls zb_sum if it has decided it has something
hypergeometric and is just about to return the piecewise representation.

We call zb_sum with J = 1 and J = 2 only to make it less likely the sum
gets caught in zb_sum for a long time. This still allows for a great
deal of sums, 'most' of the ones of interest that can be done are
liable to be caught just with J = 1, maybe.

Now for example summation(binomial(n, k) + 1, (k, 0, n)) will return
2**n + n + 1.

Will see if this breaks anything and if not have a clean up and check
everything is fine.