Tidy up zeilberger.py and change summation
Added try-catch to zb_sum as it seemed the neatest way to deal with
G_a or G_b not being defined, or rsolve recieving too complicated
an expression.

Summation used to call zb_sum with J=1 and J=2, but this can lead to
some sums taking quite a lot of time while zb_recur tries to solve the
J=2 case, and while often a recurrence is produced, rsolve is unable
to solve it so ultimately None is returned anyway.