core: Fixes #14704 - OverflowError
This patch will fix the Overflow Error encountered when the value for n is too large, such  that Python is not able to convert it to a float in the expression exp = _log(y,2)/n. It can be fixed simply by changing the expression to exp = int(_log(y,2))/n. Thanks @jksuom for your inputs!