Disable mpmath using Sage in Travis outside of the Sage tests
Mpmath using the Sage backend causes issues throughout SymPy, which are
generally bugs with upstream Sage, mpmath, and/or gmpy2 (see #14703). Also we
would really need to be running the full test suite with Sage installed to be
catching all of these (TEST_OPT_DEPENDENCY where the Sage tests are run only
runs a subset of the test suite).