Use os.path.normcase instead of sys_normcase in runtests
For whatever reason, os.path.normcase returns the string unchanged on macOS,
even though the filesystem is case insensitive.

This causes inspect.getmodule to fail on the test files when the tests are
running, because it uses sys.normcase and doesn't recognize the lowercase path
as the same as the normal path, which in turn causes a test I am writing (next
commit) to fail.