Raise an exception if sympify(safe=False) is called from SymPy library code
Since the threat model of sympify(safe=True) is that the locals dictionary is
the default, i.e., populated from 'from sympy import *', if a library function
calls sympify with safe=False, it could produce a way to work around the
safety checks (with sympify('library_func(unsafe_input)')).

This uses the inspect module to test this, but bails if it fails, so even
though it uses some hackery it shouldn't cause any problems.

This check has not been added to parse_expr yet (this is more difficult
because sympify itself calls parse_expr, so we'd have to guard against that).

It also is not possible to add a test for safe=False with this because the
tests themselves are library code (assuming they are imported).