Change the str printer of Interval to not use the class constructors
The class constructors are nice, but unfortunately, the safe sympify parser
does not allow attribute lookup, so to keep the str/srepr form supporting
sympification, we print it using keyword arguments instead, like Interval(0,
1, left_open=True, right_open=True).

This does not affect the class definitions themselves, just the string
printing. It also does not affect the pretty printers.

For now, intervals with +/- infinity print the keyword argument for that side
as open, even though integrals with infinities are always open by default.
This seems more self-documenting, although it is a bit more verbose.