Make ordinals objects not singletons
This code hasn't been in a release yet, so there are no compatibility issues.
In general, things should not be put on the singleton registry unless they are
going to be used often and the ability to do is comparison or the memory
savings are worth it. It is not necessary to make an object the only instance
of its class, since == works just fine with two instances. For example (after
this change):

>>> OrdinalOmega() == OrdinalOmega() == omega
True