More effective solve_poly_system()
The function solve_poly_system() finds solutions to a system of polynomial
equations. This is implemented using the elimination and backsubstitution:
we construct the groebner basis w.r.t. the lexicographic ordering and then we
recursively eliminate variables, starting with the univariate polynomial in
the basis. In the previous implementation, the groebner basis was computed
from scratch at each step of elimination. This is not necessary and it makes
the the solve_poly_system() very ineffective, since computation of groebner
basis can be very expensive(). In this commit, this function is rewritten so
that the groebner basis is computed only once at the beginning.