Remove automatic wrapping of input arguments for lambdify with numpy
The automatic wrapping of input arguments was to fix the behaviour of
Piecewise expressions when passed to lambdify with the numpy module
(issue #11306).  However this resulted in performance regressions for
all code using lambdify with numpy (issue #14671).

This commit removes the automatic wrapping code, and updates the lambdify
documentation to explain how to handle Piecewise with numpy.