Make zb_sum not require assumptions
Now zb_sum does not require assumptions to compute a sum, instead
it will pretend it has those assumptions, then in summation we will
return a piecewise function. This makes things slightly confusing to
read, we create a new symbol which plays the role of an old symbol
with additional assumptions, and have to substitute this in everywhere.