Fix integers returned by parametric_log_deriv_heu.
There is a typo in the pseudocode on p.249 of Bronstein's book.
The integers M and N are defined twice, in two different ways.
Only the latter definition agrees with the theory.

The returned values are also expected to be of integer type, not
polynomials, as they are to be compared with other integers.

The current version of is_log_deriv_k_t_radical_in_field returns
only two parameters.