limit(sqrt(x)*cos(x - x**2) / (x + 1), x, oo) causes recursion error
Fix issue 14574

limit(sqrt(x)*cos(x - x**2) / (x + 1), x, oo) causes recursion error. I noticed
that heuristics() in limits.py was not able to find the result.
After a change of variable, i.e. x -> 1/x, heuristics() calculates the limits
of each of the three factors and put them in the list r. At the end,
r = [0, oo, AccumBounds(-1,1)] and in this way the code rv = e.func(*r) returns
S.Nan.

My idea is to separate the factors of type AccumBounds, collected in r2, from the others,
collected in e2. Then calculate the limit for the expression obtained by the product
of elements of e2 and finally try to resolve the original limit.

A test was added.