Get the release notes from the pull request description automatically using a bot
The bot (https://github.com/sympy/sympy-bot) automatically checks every pull
request to make sure there is a valid release notes entry in the description.
The release notes entry should placed between the `<!-- BEGIN RELEASE NOTES
-->` lines and be formatted like

```
* solvers
  * improved logarithmic solving
```

In other words, each release note entry should have a heading with the
submodule that was changed, or `other`, and have a bulleted list of the
changes.

I have written a guide on how to write release notes at
https://github.com/sympy/sympy/wiki/Writing-Release-Notes.

Whenever a new pull request is made, or the description is edited, the bot
will automatically check the release notes to see if they are formatted
correctly.

When the pull request is merged, the bot will automatically take the parsed
release notes and add them to the release notes on the wiki.

Any suggestions on the wording here or in the bot are welcome.