Mul(AccumBounds(-1, 1), 0) gives AccumBounds(-1, 1)
Fix issue 13159

Debugging, I noticed that flatten() in Mul behaves in a non completely
correct way. Indeed in flatten(), seq is equal to [AccumBounds(-1, 1), 0] and at line 181 the
if statement b.is_Rational is True; hence the order of the list seq is reversed in the
variable a and b, but not in the list itself.

Added a test.