Fix in `JointRV`
Fix in `JointRV` to order indexed symbols extracted by their index,
instead of using the set as it is. Also replaced symbols by indexed
symbols in `marginal_distribution`. `Indexed` are not used directly and
instead replace `Symbol` towards the end of the function because
`integrate` raised error in some cases when the variable of integration
was of type `Indexed`.