Fix issue 13159
Changing the line 265 of mul.py in the following way:
elif coeff.is_Number: # it could be zoo (original line)
elif coeff.is_Number or isinstance(coeff, AccumBounds): # it could be zoo (modified)

the code

                    coeff *= o
                    if coeff is S.NaN:
                        # we know for sure the result will be nan
                        return [S.NaN], [], None

is executed also in the case coeff is AccumBounds(-1, 1) (or similar) and,
in case o is 0, then coeff becomes 0.

A test is added.