Integrate gives a wrong result
The PR fixes issue 14782.
integrate(sqrt(-x**2 + 1)*(-x**2 + x), [x, -1, 1]) gives 0 instead -pi/8.
The antiderivative works correctly, but it returns a function containing
a Piecewise expression, defined in the open interval (-1, 1).
The function _eval_interval() does not work correctly, since the antiderivative
is not a Piecewise expression.
Hence, separating the terms of type Piecewise and using the _eval_interval() for
Piecewise() solves the issue.

Added a test.
Unfortunately, a test in test_ode.py fails due to the fact that a minus sign compares
inside Piecewise intead that outside. So I modified such a test, without changing the
final result.