meijerint_inversion: exp: don't return unevaluated in Piecewise
If `cond` is True, it would be discarded anyway.  If `cond` is an
expression (an Equality) then what we really want to say is that
the transform does not exist.  We don't currently have a way to do
that precisely but a Piecewise we no default `True` subfunction
should be mathematically correct (in that the result is undefined
when the condition is false).  Leave a comment in the code to note
these are the only two options for `cond` (in contrast with the
later code in this function).