UndefinedFunctions with _imp_ can evalf
The lambdify machinary has `implemented_function` to add implementation
to an UndefinedFunction: not quite sure why we need this but since its
there let's not break it.  If an UndefinedFunction has an `._imp_`
method, than we should let it evalf.  But we still shouldn't look it up
in mpmath so add a test for an UndefinedFunction with a suspicious name
and a `._imp_`.