doslve: further cleanups to `function=True`
In two cases, it would be wrong to replace them with `cls=Function` because
the whole point is to sub in Symbols for functions for later manipulation.
I suppose they could stay as `function=True` but I think that's more likely
to confuse future maintainers than help them remember the semantics of these
variables.

The other case in ` _linear_2eq_order2_type10` is untested.  It looks like
they should be `cls=Function` so I made that change.  Filed #15097 asking for
a new test.