Store the first argument of MatrixSymbol as a Symbol
MatrixSymbol.name is still a string.

This is a backwards incompatible change, but one that has needed to be changed
for some time. The issue is that various places in SymPy assume that all args
are SymPy types, and hence fail if they are not. This is exacerbated by
functions that rebuild expressions after sympifying the args. It was possible,
for instance, with something like

A = MatrixSymbol('A', n, n)
B = MatrixSymbol('B', n, n)
factor(A*B)

to create an expression that would fail in the printers, because it replaced
MatrixSymbol('A', n, n) with MatrixSymbol(Symbol('A'), n, n). This is because
factor (actually factor_nc) does effectively

expr = sympify(expr)
return expr.func([factor_nc(i) for i in expr.args])

I have not yet gone through the issue tracker to try to find things that this
might fix, nor have I run the test suite beyond sympy/matrices, which had no
failures. The only test that I've added thus far is to un-XFAIL the test_args
test for MatrixSymbol.