Correctly handle defining a Singleton class with the same name twice
Without this, there are doctest failures from the Singleton doctests when its
run twice in the same Python session (doctest('singleton', subprocess=False)).

There are still some side effects from running the tests or doctests with
subprocess=False, in that the test singleton classes are registered on the S
object.

Fixes a failure from #15149.