use max_cosets instead of changing CosetTable.coset_table_max_limit
Before introduction of the keyword `max_cosets`, the only way to
control the maximum number of cosets in a coset enumeration was
by changing the class attribute `coset_table_max_limit`. If not
restored to the original value, the change in attribute affects
all subsequent coset enumerations causing doctests to fail if run
twice in the same process. This commit changes the tests to use
`max_cosets` instead but also modifies the code of CosetTable to
allow the attribute change to be preserved even if `max_cosets`
is used after the change