Move the super() call to the end of FunctionClass.__init__
The super() call calls BasicMeta.__init__, which adds the class to a set of
classes (sympy.core.core.all_classes). This computes the hash of the class.
However, the hash of a Function depends on its nargs (class_key()), meaning if
it is set dynamically, the hash will change and the class won't be found in
the all_classes set any more, breaking sympify(obj).

This happens for things like Function('f', eval=lambda self, t: ...) where
nargs is determined from the eval lambda.

Fixes #15170.