linear_coeffs added to solveset
In addition, nonlinearity in Eq which would
cancel by subtraction is now detected in
linear_eq_to_matrix.

Note: the requirement that symbols be Symbols
is made to avoid problems like finding the
coefficients of x in `x + f(x)`. `as_coeff_mul(x)`
will treat both terms as x-dependent and will not
return f(x) as a constant and 1 as the coefficient.