differentiation modifications
Notes:

Objects define their own diff if they want to take control of how
differentiation is handled: diff(foo) will call foo.diff() if it is
defined. There is no need for `_eval_diff` routines (though the
definitions of diff could be renamed _diff like subs. In fact, it
might be better to do so: canonicalization and sanitation of input
happens in diff and then _diff is called once something significant
needs to take place.

Here is the subs model

.subs - handles sanitizing of input and calls _subs if there is
        something to be done
    ._subs - handles traversal of arguments
        ._eval_subs - handles substitution of old->new for an object

diff doesn't have as clearly delineated operations.
It would be good to split the sanitization of input and maybe trivial zero
detection from the process of calculation the sequence of derivatives.
Methods that define diff methods should probably be defining _diff
methods.

.diff - handles sanitizing and trivial zero detection and
        calls _eval_diff routines
    ._eval_diff - called from diff to compute d_/dx
    ._eval_diff_n_times - called from diff to compute d^n_/dx^n
.fdiff - computes the 1st derivative of a functions ith argument