broaden dummy_eq capabilities
Subs has been given a method of computing hashable content
so expressions that are trivially different compare the same.
The same has not been done for ExprWithLimits. To aid in
their comparison (and in the general representation of objects
with structural dummy symbols) the following modifications have
been made:

* core
  * `dummy_eq` now compares expressions with canonical dummy variables
  * `canonical_variables` now returns non-clashing Symbols
  * `dummy_eq` allows keyword `canonical` to replace structural dummy variables
     with canonical symbols (u, v, w, x, y, z, iota_0, iota_1, ...)
* concrete
  * dummy_variables returns the structural dummy variables of the expr

Any object that has structural dummies should define a `dummy_variables`
property that returns an iterable of its dummy variables.