Fixed Beam.apply_load() wrong result for high order loads with ends
Fixes #15301

We were incorrectly calculating the subtractive load if the user
supplied an end distance for a distributed load of order 2 and higher.
To solve this a Taylor series is used to compute the proper subtractive
function. @faradai2008 figured out the exact form of the series and it
has now been implemented with a loop summation. Tests have been added
for basic cases and prior incorrect tests have been modified.