Pass the latex_mode flag of init_printing() through to the latex formatter
This also has the effect of making the default mode 'equation*', instead of
'plain'. This mode was already being used for the png formatter, and it seems
more appropriate for the notebook, as it does large equation formatting. For
instance, integration limits are rendered above and below the integral sign
with equation*, whereas with plain they are rendered to the right of the
integral sign.