Reverted changes to code printer due to performance issues.
Fixes #15377

The C code printer was changed in #13692 to support the Mod function by
including a set of lambda statements in known_function. This required
many more recursions in the printer, slowing things considerably for
function printing. Some of that change has been reverted and the printer
now uses a _print_Mod to handle the logic.