Reverts backwards incompatible changes to Matrix.diff()
PR #13641 introduced a new feature: differentiating Matrices by Matrices
and Arrays. To do this, three existing tests were changed, breaking
backwards compatibility. The changes to the code caused
MutableDenseMatrix.diff() to return an ImmutableDenseMatrix. This breaks
user code that relies on the mutability of the matrices. In particular,
it has large effects to the sympy.physics.mechanics code that relies
heavily on matrix differentiation. This PR reverts the three changed
tests, reverts the change to MatrixCalculus.diff() that causes the type
change, and moves all subsequent tests related to differentiating
Matrices with respect to Matrices/Array/etc into an XFAIL test, as they
are now broken.