Reverts backwards incompatible changes to theano_function()
PR #14949 introduced a backwards incompatible change to the public API
of the theano_function() function. The change allowed the output
expressions to be passed in without being encased in a sequence, but did
not retain the prior behavior. The expressed reasons for the change were
to allow a user to type `my_expr` instead of `[my_expr]` and to have a
similar call signature with `theano.function()`. Neither of these
warrant a necessary deprecation. In particular, commit
a833af605791c38518d133fd98d7d0e99190ff97 is reverted.

Secondly, this change inadvertently broke (seemingly) untested behavior
if the output expressions are matrices. Prior to the change a numpy
array of the same dimensions of the matrix would be returned. The change
caused lists of arrays to be returned. This returns that behavior and
adds a regression test.