Restore a non-Hermitian version of the Cholesky and LDL decompositions
The standard Cholesky decomposition has the form A = L*L.H and as such
applies only to Hermitian positive-definite matrices. However, in
practice it is sometimes convenient to use the form A = L*L.T where L
may have complex entries (whether or not A has them). Since L is lower
triangular, this "non-Hermitian Cholesky decomposition" can be used to
solve linear systems just as well. In some situations, such as the SymPy
benchmark for `cholesky_solve`, the non-Hermitian decomposition is
preferable: the given symbolic matrix is symmetric but there is no way
to know if it is positive-definite. Now the `cholesky_solve` method uses
the non-Hermitian decomposition whenever the matrix is symmetric.

Same changes for LDL decomposition. More tests and examples added.