Fixes #15548: Union Supports Python set
Union Calls _sympify when an argument is a python set, except when
any element of that set is a sympy Set. This is because _sympify gives
untinded result when an element of set is a Set.
Example
>>> Union({1},{2},{3})
{1,2,3}
>>> type(Union({1},{2},{3}))
<class 'sympy.sets.sets.FiniteSet'>

However, This does not work.
>>> Union({1,Interval(0,1)})
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "sympy/sets/sets.py", line 1036, in __new__
    return S.EmptySet
  File "sympy/sets/sets.py", line 1034, in flatten
    # Union of no sets is EmptySet
  File "sympy/sets/sets.py", line 1034, in flatten
    # Union of no sets is EmptySet
  File "sympy/sets/sets.py", line 1035, in flatten
    if len(args) == 0:
TypeError: Input must be Sets or iterables of Sets

Adds Test in test_union() for checking this functionality.