Add support for overdetermined systems to LUsolve
Some overdetermined linear systems (with more equations than unknowns)
are nonetheless consistent, i.e., have a solution. The present LUsolve
algorithm requires only minimal modifications to handle such systems.

Indeed, after row permutation and forward substitution, the redundant
rows of the coefficient matrix are at the bottom and have zero entries.
So the consistency check amounts to verifying that the right hand side
is also zero in those rows. If this is the case, the trivial equations
"0 = 0" are ignored, and a solution is obtained.

In contrast, the present LUsolve algorithm is not well fit for
underdetermined systems, for which it can give absurd results:
`Matrix([[0, 1]]).LUsolve(Matrix([1]))` returns `[[zoo]]` instead of
`[[0], [1]]` or `[[whatever], [1]]`.  So I added NotImplementedError for
the underdetermined case.