Union does not flatten. Nested lists are converted to sets inside set
While using the Union constructor, the sets can be inside a list, an iterable or unpacked.
However, list inside the outermost list or iterable are taken to mean as another set.

>>> Union([{1},{2}])
{1, 2}
>>> Union([{1},[{2}]])
{1, {2}}

Also, if more than one arguments are passed unpacked, those containing a list are taken
to be a set.

>>> Union({1},[{2}])
{1, {2}}

>>> Union(FiniteSet(1),[FiniteSet(2),[FiniteSet(3)]])
{1,{2},{{3}}}

Changes a test that used old convention.