Request an integer number of terms from _eval_nseries
The method `Expr.series` predicts the number of terms to be requested
from `_eval_nseries` with the linear extrapolation formula
`ndo = n + (n - ngot)*more/(newn - ngot)`.
But this quantity may be fractional, which does not fit "the number of terms"
concept. Replaced with
`ndo = n + ceiling((n - ngot)*more/(newn - ngot))`
that is, rounding up.