Implement Idx.name property
Added "name" property, which is usually the name of the symbol used as
indexed label. If an Idx is created from an expression like
`Idx(2*k+1)`, then the string representation of that expression is used.

Rationale: Idx class is considered symbol-like (is_symbol set on True)
but it does not have a property "name" like other symbols and
symbol-like objects (Indexed, IndexedBase, MatrixSymbol). This inconsistency
causes errors because symbols are expected to have names.