Correct check for whether nested powers (x**a)**b can be combined
Currently `Pow.subs` decides whether (x**a)**b can be combined on the
basis of `Pow._eval_power(x**a, b)` not returning None. However, the
return value can be Mul, as in `(x**2)**(3/2)` evaluating to
`x**2*Abs(x)` when x is real. This does not mean the exponents could be
combined. The correct check is `isinstance(..., (Pow, exp))` instead of
`is not None`.