Account for multiple "extra" types in IntegralTransform.doit
The method IntegralTransform.doit is common to several types of
transforms, which present conditions for convergence in different ways:
just a condition, or (abscissa of convergence, condition) for Laplace
transform, or ((pair of abscissas), condition) for Mellin Transform. The
method needed to be more aware of this distinction by checking the kinds
of objects it's handling.