Simpler expansion for Fresnel integrals at infinity
The expansion implemented in `fresnels._eval_aseries` is taken from a
paper which derives an asymptotic for Fresnel integrals in a sector of
the complex plane. As a result, the terms cos(z) and sin(z) are included
as factors of O() in the expansion, because they are not necessarily
small. However, `series(fresnels(x), x, oo)` is explicitly a
representation on the real line, which means cos and sin can be absorbed
as O(1). Doing so results in a more usable series, and solves several
problems:

Incorrect limits at oo:
```
>>> limit(fresnels(x), x, oo)
0
>>> limit(3 + fresnels(x), x, oo)
0
```
Inconsistency between the series for fresnels and 2*fresnels:
```
>>> series(fresnels(x), x, oo)
(-1/(pi**2*x**3) + O(x**(-6), (x, oo)))*sin(pi*x**2/2) + (3/(pi**3*x**5) - 1/(pi*x) + O(x**(-6), (x, oo)))*cos(pi*x**2/2) + 1/2
>>> series(2*fresnels(x), x, oo)
6*cos(pi*x**2/2)/(pi**3*x**5) - 2*sin(pi*x**2/2)/(pi**2*x**3) - 2*cos(pi*x**2/2)/(pi*x) + 1 + O(x**(-6), (x, oo))
```
(The former is hardcoded, the latter is computed by SymPy.)

In addition, the series at -oo is now handled as well.