exptrigsimp: Avoid use of as_powers_dict on noncommutatives
A slightly hackish solution to the first part of #15551. There is probably
room for more improvement by exploiting what commutativity is available.
exp(a)*a*exp(a) no longer simplifies to exp(2*a)*a when a is noncommutative.
Still, it is better to avoid the wrong answers that motivated this change.