Make the substitution of symbols consistent with powers of symbols
The `subs` method now applies the same logic to symbols as to their
powers, by calling `Pow._eval_subs`. Old behavior:
```
>>> (log(x)**2/x).subs(1/x, u)
u*log(x)**2
```
New behavior:
```
>>> (log(x)**2/x).subs(1/x, u)
u*log(1/u)**2
```
This uncovered some issues with `manualintegrate`: it is prone to
infinite recursion, which was previously prevented by the failure of subs.
This is corrected as well.