Remove undocumented "sequence" flag from Limit.doit()
Per the discussion in #14327, the undocumented `sequence` parameter of
`Limit.doit()` is removed. The docstring of limit now refers to
limit_seq for computing limits of sequences. The remaining keywords
arguments of `Limit.doit()` and limit are described in docstrings. Some
tests in test_limits, which relied on the sequential behavior, are moved
to test_limitseq. A test for `limit(exp(I*x)*sin(pi*x), x, oo)` is
added -- it incorrectly evaluates to 0 in current master.

Additional changes post #14378 are minor: `limit_seq` is used instead of
`limit` in the limit comparison and bounded-times-convergent tests for
series convergence.