Use appropriate NumPy methods for matrix inverses and powers
Currently NumPy printer prints `A**2` as such even if A is a
MatrixSymbol, resulting in elementwise squaring.  It should use
`numpy.linalg.matrix_power` instead. Similarly for matrix inverse.
Changes are tested with lambdify as well.

In addition, a string representation of `A**(-1)` is corrected:
previously `str(A**(-1))` returned `A^-1`, in contrast with
`str(A**(-2))` returning `A**(-2)`.