Adjust the logic of Abs->Piecewise rewrite in integrate
According to a comment in integrals.py, "some improper integrals are
better off with Abs", that is, with Meijer G method rather than
Piecewise logic. This motivated the condition `all(x.is_real and
x.is_finite for x in xab[1:])` for automatic Abs->Piecewise rewrite. But
this condition leaves Abs in when `x.is_finite` is None, which offers no
advantage because the integral is not actually known to be improper.
Instead it should be `all(x.is_real and not(x.is_infinite) for x in
xab[1:])`.