lambdify: Correct the logic of namespace generation in _import
The signature `_import(module, reload="False")` has a typo: the
parameter `reload` is used as a bool and "False" is a truthy string.
Correcting this brings up another issue: the function uses `namespace !=
namespace_default` as an indication that "The namespace was already
generated" but this isn't so because, for example, the initial values of
`NUMPY` and `NUMPY_DEFAULT` are already different (contrary to the
purpose of having a default namespace).  To avoid such discrepancies,
NUMPY should be created as a copy of `NUMPY_DEFAULT`, and similarly for
others.