Improve logarithmic substitution in manualintegrate
When the substitution `subs(log(x), u)` is made, SymPy does not replace
`x` with `exp(u)`, nor does it replace `x**a` with `exp(a*u)`. This is
problematic for manual integration which needs to completely replace the
variable of integration by a new one.

To solve this, a helper function `manual_subs` is added to
manualintegrate. Similarly to the existing `manual_diff`, it is a
wrapper that optimizes the outcome of substitution for the purpose of
manualintegrate. As a result, a bunch of new integrals get evaluated
with `integrate`.