Improve the handling of oscillatory sequences in limit_seq
Oscillatory sequences are handled by considering odd and even indices
separately (this is more efficient than replacing n by `2*n` or `2*n+1`
as was previously done). Also, the presence of sin or cos in the formula
now qualifies a sequence as oscillatory.

When the limits along even and odd indices exist but are different,
AccumBounds is returned instead of None.

For some sequences, the absolute value can be handled by `limit_seq`
while the original sequence cannot (for example, `(I/2)**n`). If the
limit of absolute value is 0, the limit of the original sequence is also
0.