integrals: Add 'heurisch' flag to integrate()
The function `integrate` has boolean flags risch, meijerg, and manual
which can be used to disable a particular method (False), or to use only
that method (True). This adds such a flag for the other general
integration method, "heurisch" (parallel / heuristic Risch algorithm).
This method is known to sometimes take extremely long time (practically
hanging the computation), so it may be desirable to disable it in some
contexts. Also, these flags make it easier to debug `integrate`.