minpoly-compose: preserve exponent of -1
`_minpoly_compose` computes the minimal polynomial of a product
by first combining all factors `b**e` where `b` and `e` are
rational. Each exponent `e` is written as `n/d` where `d` is the
least common multiple of all exponents. Then the product of the
powers `b**n` is computed and finally its `d`th root is taken.
This may give a result with wrong sign if `b == -1` and `n` is
even. That is what happens with one term of the expression in
issue 14831:

    >>> minimal_polynomial(-2*sqrt(2)*sqrt(12*sqrt(2) + 17), x)
    x**2 - 16*x - 8

This is the minimal polynomial of `2*sqrt(2)*sqrt(12*sqrt(2) + 17)`.
The correct result is `x**2 + 16*x - 8`.

This patch deals with powers of -1 separately taking care of
preserving the sign.