Request people to manually backreference pull requests in the issue
That is, in the pull request template.

The problem is that even though GitHub does automatically add cross references
when the issue is referenced, it does so in a way that most new people tend to
miss. An actual comment is more likely to be noticed, and is better practice
anyway.