Remove the automatic reordering of the "notes" and "examples" sections in Sphinx
There is no special formatting applied to these sections (we do not use the
plot feature of numpydoc), so there is no reason to treat them as special
headers. It's annoying to have these sections in a specific place in the
docstring and for Sphinx to move them around. For example, if you put "notes"
at the end, without this patch, it would move it to the top.

It still does reorder other sections, such as see also and references, but
these have special formatting applied to them, so it's less of an issue.
Potentially moving to napoleon would fix this (I haven't 100% confirmed). See
issue #15375.