Numbers: Removed the redundant checks of igcd method and simplified it
In igcd method various redundant checks are present such as in line 185 a is surely not equal to 1.
Also in line 192, b != 1 as we have checked that earlier in line 179. Many other such redundancies have been removed.
I have done this as advised by @oscargus and ran all the checks.