PermutationGroup.is_normal: Fixed check for identity group, revised tests
The implementation that uses PermutationGroup.coset_factor to check if a
permutation p = g*h*g^{-1} (g \in G, H \in H, H is tested to be a normal
subgroup of G) is again in H (that is the PermutationGroup instance on
which the method is called) has the problem that if h is the identity
permutation the check fails. As a consequence, the result of the
is_normal check is incorrect. This contribution extends to check to
recognize the identity permutation.
Furthermore, the subgroup property was never checked which also led to
incorrect results (i.e. a group of larger order was identified to be a
normal subgroup of a group of lower order).
Tests are extended and an incorrect test is removed.