Disallow noncommutative scalars in MatMul
The problem is that some algorithms try to replace matrices with
noncommutative Dummys. But this only works if it can keep them in the right
order, which presently it doesn't, as it separates out the scalar and matrix
parts. Since the current behavior can lead to wrong results, e.g., with
factor_nc(), it is better to explicitly disallow it.

This breaks factor_nc() from working on matrix expressions. However, it only
worked correctly in some cases. For example,
`factor_nc(B**(-1)*(A**(-1)*B**(-1) - A**(-1)*C*B**(-1))**(-1)*A**(-1))`
returned `B**(-1)*(((1 - C)*B**(-1))*A**(-1))**(-1)*A**(-1)`, which not only
reordered the terms incorrectly, but used 1 instead of I.