Add UndefinedMatrixShapeSymbol class
This can be used as a symbolic shape in a matrix expression, and will not
result in errors when compared to a different shape. For instance,

    _n = UndefinedMatrixShapeSymbol('_n')
    k = Symbol('k')
    A = MatrixSymbol("A", _n, _n)
    B = MatrixSymbol("B", k, k)
    A + B # does not result in an error

In particular, MatAdd() and MatMul() with no arguments now return a ZeroMatrix
and Identity with Dummy UndefinedMatrixShapeSymbol, rather than scalar 0 and
1.

Also makes it so that is_square returns None if one of the arguments has an
UndefinedMatrixShapeSymbol.

See the discussion at
https://github.com/sympy/sympy/pull/15948#issuecomment-463432718.