Add GenericZeroMatrix and GenericIdentity class in matrix expressions
These are generic in the sense that they have no shape (.shape raises a
TypeError). They are returned by MatAdd() and MatMul() with no arguments.

In order to avoid the TypeErrors being an issue, MatAdd and MatMul remove them
unconditionally in their constructors (normally they are completely
unevaluated without calling doit()).

This is an alternative to PR #15986. See the discussion there and at
https://github.com/sympy/sympy/pull/15948#issuecomment-463432718.