Replaced redundant checks
In `test_arit.py` some tests of type `assert foo = foo` were present which
were not doing really much. The key idea to keep these tests was to check if
the expression remains unchanged but `assert foo == foo` does not ensures that.
In this change, `unchanged` function is used to check if the expressions
remain unchanged or not.