misc: add strlines
When testing the output of the strlines function, the
rawlines function was used to prepare the output to
use in the test. Normally, the way this would be used, however,
would be to have

Given: func(s) -> longstring
print(strlines(func(s)))
<copy what was printed>
assert func(s) == <paste>

A quoting issue was fixed for rawlines, too. Instead of adding
a quote string the repr form is used.